update treeview demos
