+++ docs/src/components/productX/XTreeViewDemo.tsx
@@ -1,260 +1,274 @@
 import * as React from 'react';
 import clsx from 'clsx';
+import { animated, useSpring } from '@react-spring/web';
 import { styled } from '@mui/material/styles';
-import Box from '@mui/material/Box';
-import Paper from '@mui/material/Paper';
-import { SimpleTreeView } from '@mui/x-tree-view/SimpleTreeView';
-import {
-  TreeItem as MuiTreeItem,
-  useTreeItemState,
-  TreeItemProps,
-  TreeItemContentProps,
-} from '@mui/x-tree-view/TreeItem';
-import Typography from '@mui/material/Typography';
-import FolderRounded from '@mui/icons-material/FolderRounded';
-import FolderOpenRounded from '@mui/icons-material/FolderOpenRounded';
-import PhotoOutlined from '@mui/icons-material/PhotoOutlined';
-import PictureAsPdfOutlined from '@mui/icons-material/PictureAsPdfOutlined';
-import VideocamOutlined from '@mui/icons-material/VideocamOutlined';
-import FourKOutlined from '@mui/icons-material/FourKOutlined';
+import { TransitionProps } from '@mui/material/transitions';
 import { HighlightedCode } from '@mui/docs/HighlightedCode';
 import Frame from 'docs/src/components/action/Frame';
-import TreeViewDemo from './TreeViewDemo';
+import Box from '@mui/material/Box';
+import Collapse from '@mui/material/Collapse';
+import Typography from '@mui/material/Typography';
+import Paper from '@mui/material/Paper';
+import FolderRounded from '@mui/icons-material/FolderRounded';
+import FolderOpenRounded from '@mui/icons-material/FolderOpenRounded';
+import ImageIcon from '@mui/icons-material/Image';
+import PictureAsPdfIcon from '@mui/icons-material/PictureAsPdf';
+import VideoCameraBackIcon from '@mui/icons-material/VideoCameraBack';
+import { RichTreeView } from '@mui/x-tree-view/RichTreeView';
+import { treeItemClasses } from '@mui/x-tree-view/TreeItem';
+import {
+  unstable_useTreeItem2 as useTreeItem2,
+  UseTreeItem2Parameters,
+} from '@mui/x-tree-view/useTreeItem2';
+import {
+  TreeItem2Content,
+  TreeItem2IconContainer,
+  TreeItem2Label,
+  TreeItem2Root,
+} from '@mui/x-tree-view/TreeItem2';
+import { TreeItem2Icon } from '@mui/x-tree-view/TreeItem2Icon';
+import { TreeItem2Provider } from '@mui/x-tree-view/TreeItem2Provider';
+import { TreeViewBaseItem } from '@mui/x-tree-view/models';
 
-const CustomContent = React.forwardRef(function CustomContent(
-  props: TreeItemContentProps & { lastNestedChild?: boolean },
-  ref,
-) {
-  const {
-    lastNestedChild,
-    classes,
-    className,
-    label,
-    itemId,
-    icon: iconProp,
-    expansionIcon,
-    displayIcon,
-  } = props;
+type FileType = 'image' | 'pdf' | 'video' | 'folder';
 
-  const {
-    disabled,
-    expanded,
-    selected,
-    focused,
-    handleExpansion,
-    handleSelection,
-    preventSelection,
-  } = useTreeItemState(itemId);
+type ExtendedTreeItemProps = {
+  fileType?: FileType;
+  id: string;
+  label: string;
+};
 
-  const icon = iconProp || expansionIcon || displayIcon;
+const ITEMS: TreeViewBaseItem<ExtendedTreeItemProps>[] = [
+  {
+    id: '1',
+    label: 'Drive',
+    children: [
+      {
+        id: '1.1',
+        label: 'Backup',
+        children: [
+          { id: '1.1.1', label: 'Jan 2023.pdf', fileType: 'pdf' },
+          { id: '1.1.2', label: 'Feb 2023.pdf', fileType: 'pdf' },
+          { id: '1.1.3', label: 'Mar 2023.pdf', fileType: 'pdf' },
+        ],
+      },
+      {
+        id: '1.2',
+        label: 'Photos',
+        children: [
+          { id: '1.2.1', label: 'family.jpeg', fileType: 'image' },
+          { id: '1.2.2', label: 'my_dog.png', fileType: 'image' },
+        ],
+      },
+    ],
+  },
+  {
+    id: '2',
+    label: 'Favorites',
+    children: [
+      {
+        id: '2.1',
+        label: 'MUI_retreat_photo.jpg',
+        fileType: 'image',
+      },
+      {
+        id: '2.2',
+        label: 'v7_secrets.mkv',
+        fileType: 'video',
+      },
+      {
+        id: '2.3',
+        label: 'Other pictures',
+        children: [{ id: '2.3.1', label: 'my_avatar.jpeg', fileType: 'image' }],
+      },
+    ],
+  },
+];
 
-  const handleMouseDown = (event: React.MouseEvent<HTMLDivElement>) => {
-    preventSelection(event);
-  };
+declare module 'react' {
+  interface CSSProperties {
+    '--tree-view-color'?: string;
+    '--tree-view-bg-color'?: string;
+  }
+}
 
-  const handleExpansionClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    handleExpansion(event);
-    handleSelection(event);
-  };
+const StyledTreeItemRoot = styled(TreeItem2Root)(({ theme }) => ({
+  color: theme.palette.mode === 'light' ? theme.palette.grey[800] : theme.palette.grey[400],
+  position: 'relative',
+  [`& .${treeItemClasses.groupTransition}`]: {
+    marginLeft: theme.spacing(3.5),
+  },
+})) as unknown as typeof TreeItem2Root;
 
-  const handleSelectionClick = (event: React.MouseEvent<HTMLDivElement>) => {
-    handleSelection(event);
-  };
+const CustomTreeItemContent = styled(TreeItem2Content)(({ theme }) => ({
+  borderRadius: theme.spacing(0.7),
+  marginBottom: theme.spacing(0.5),
+  marginTop: theme.spacing(0.5),
+  padding: theme.spacing(0.5),
+  [`& .${treeItemClasses.iconContainer}`]: {
+    marginRight: theme.spacing(1),
+  },
+  '&.Mui-expanded&::before': {
+    content: '""',
+    display: 'block',
+    position: 'absolute',
+    left: '16px',
+    top: '44px',
+    height: 'calc(100% - 48px)',
+    width: '1.5px',
+    backgroundColor:
+      theme.palette.mode === 'light'
+        ? (theme.vars || theme).palette.grey[100]
+        : (theme.vars || theme).palette.primaryDark[700],
+  },
+}));
 
-  const renderExtension = () => {
-    if (typeof label !== 'string') {
-      return label;
-    }
-    const extension = (label || '').split('.').slice(-1)[0];
-    if (extension.match(/(jpg|jpeg|png)/)) {
-      return <PhotoOutlined />;
-    }
-    if (extension === 'pdf') {
-      return <PictureAsPdfOutlined />;
-    }
-    if (extension === 'mp4') {
-      return <VideocamOutlined />;
-    }
-    if (extension === 'mkv') {
-      return <FourKOutlined />;
-    }
-    return (
-      <Box
-        sx={{
-          width: 6,
-          height: 6,
-          borderRadius: '50%',
-          bgcolor: 'warning.main',
-          display: 'inline-block',
-          verticalAlign: 'middle',
-          zIndex: 1,
-        }}
-      />
-    );
-  };
+const AnimatedCollapse = animated(Collapse);
 
+function TransitionComponent(props: TransitionProps) {
+  const style = useSpring({
+    to: {
+      opacity: props.in ? 1 : 0,
+      transform: `translate3d(0,${props.in ? 0 : 20}px,0)`,
+    },
+  });
+
+  return <AnimatedCollapse style={style} {...props} />;
+}
+
+interface CustomLabelProps {
+  children: React.ReactNode;
+  icon?: React.ElementType;
+  expandable?: boolean;
+}
+
+function CustomLabel({ icon: Icon, expandable, children, ...other }: CustomLabelProps) {
   return (
-    /* @ts-ignore -- Key event is handled by the TreeView */
-    <Box
-      className={clsx(className, classes.root, {
-        [classes.expanded]: expanded,
-        [classes.selected]: selected,
-        [classes.focused]: focused,
-        [classes.disabled]: disabled,
-        'CustomContent-last': lastNestedChild,
-      })}
-      onClick={handleExpansionClick}
-      onMouseDown={handleMouseDown}
-      ref={ref as React.Ref<HTMLButtonElement>}
+    <TreeItem2Label
+      {...other}
       sx={{
-        border: 'none',
-        borderRadius: '5px',
-        textAlign: 'left',
-        position: 'relative',
-        zIndex: 1,
-        '& svg': {
-          fontSize: 18,
-          color: itemId.startsWith('root') ? 'primary.main' : 'grey.600',
-        },
-        '&:not(.CustomContent-last)': {
-          '& svg': {
-            '&:first-of-type': {
-              fontSize: 14,
-              color: 'primary.main',
-            },
-          },
-        },
+        display: 'flex',
+        alignItems: 'center',
       }}
     >
-      {icon}
-      {lastNestedChild && renderExtension()}
-      {!lastNestedChild && (expanded ? <FolderOpenRounded /> : <FolderRounded />)}
+      {Icon && (
+        <Box
+          component={Icon}
+          className="labelIcon"
+          sx={(theme) => ({
+            mr: 1,
+            fontSize: '1rem',
+            color: expandable ? theme.palette.primary.main : theme.palette.grey[600],
+          })}
+        />
+      )}
+
       <Typography
-        onClick={handleSelectionClick}
-        component="div"
-        className={classes.label}
-        noWrap
-        sx={{
-          '&&': {
-            color: lastNestedChild ? 'text.secondary' : 'text.primary',
-            fontWeight: lastNestedChild ? 400 : 500,
-          },
-        }}
+        sx={(theme) => ({
+          fontWeight: expandable
+            ? theme.typography.fontWeightMedium
+            : theme.typography.fontWeightRegular,
+          color: expandable ? theme.palette.text.primary : theme.palette.grey[600],
+        })}
+        variant="body2"
       >
-        {label}
+        {children}
       </Typography>
-    </Box>
+    </TreeItem2Label>
+  );
+}
+
+const isExpandable = (reactChildren: React.ReactNode) => {
+  if (Array.isArray(reactChildren)) {
+    return reactChildren.length > 0 && reactChildren.some(isExpandable);
+  }
+  return Boolean(reactChildren);
+};
+
+const getIconFromFileType = (fileType: FileType) => {
+  switch (fileType) {
+    case 'image':
+      return ImageIcon;
+    case 'pdf':
+      return PictureAsPdfIcon;
+    case 'video':
+      return VideoCameraBackIcon;
+    case 'folder':
+      return FolderRounded;
+    default:
+      return FolderRounded;
+  }
+};
+
+interface CustomTreeItemProps
+  extends Omit<UseTreeItem2Parameters, 'rootRef'>,
+    Omit<React.HTMLAttributes<HTMLLIElement>, 'onFocus'> {}
+
+const CustomTreeItem = React.forwardRef(function CustomTreeItem(
+  props: CustomTreeItemProps,
+  ref: React.Ref<HTMLLIElement>,
+) {
+  const { id, itemId, label, disabled, children, ...other } = props;
+
+  const {
+    getRootProps,
+    getContentProps,
+    getIconContainerProps,
+    getLabelProps,
+    getGroupTransitionProps,
+    status,
+    publicAPI,
+  } = useTreeItem2({ id, itemId, children, label, disabled, rootRef: ref });
+
+  const item = publicAPI.getItem(itemId);
+  const expandable = isExpandable(children);
+  let icon;
+  if (expandable) {
+    if (status.expanded) {
+      icon = FolderOpenRounded;
+    } else {
+      icon = FolderRounded;
+    }
+  } else if (item.fileType) {
+    icon = getIconFromFileType(item.fileType);
+  }
+
+  return (
+    <TreeItem2Provider itemId={itemId}>
+      <StyledTreeItemRoot {...getRootProps(other)}>
+        <CustomTreeItemContent
+          {...getContentProps({
+            className: clsx('content', {
+              'Mui-expanded': status.expanded,
+              'Mui-selected': status.selected,
+              'Mui-focused': status.focused,
+              'Mui-disabled': status.disabled,
+            }),
+          })}
+        >
+          {expandable && (
+            <TreeItem2IconContainer {...getIconContainerProps()}>
+              <TreeItem2Icon status={status} />
+            </TreeItem2IconContainer>
+          )}
+
+          <CustomLabel {...getLabelProps({ icon, expandable })} />
+        </CustomTreeItemContent>
+        {children && <TransitionComponent {...getGroupTransitionProps()} />}
+      </StyledTreeItemRoot>
+    </TreeItem2Provider>
   );
 });
 
-const StyledTreeItem = styled(MuiTreeItem)(({ theme }) => [
-  {
-    paddingTop: 5,
-    '& .MuiTreeItem-content .MuiTreeItem-label': {
-      paddingLeft: theme.spacing(0.75),
-    },
-    '& .MuiTreeItem-root': {
-      position: 'relative',
-      '&::before': {
-        content: '""',
-        display: 'block',
-        position: 'absolute',
-        left: -14,
-        height: '100%',
-        width: 1.5,
-        backgroundColor: (theme.vars || theme).palette.grey[100],
-      },
-    },
-    '& .MuiTreeItem-content': {
-      padding: theme.spacing('2px', 0.5),
-    },
-    '& .MuiTreeItem-group': {
-      marginLeft: 0,
-      paddingLeft: theme.spacing(3),
-    },
-  },
-  theme.applyDarkStyles({
-    '& .MuiTreeItem-root': {
-      '&::before': {
-        backgroundColor: (theme.vars || theme).palette.primaryDark[700],
-      },
-    },
-    '& .MuiTreeItem-group': {
-      '& .MuiTreeItem-content': {
-        '&::before': {
-          backgroundColor: (theme.vars || theme).palette.primaryDark[500],
-        },
-      },
-    },
-  }),
-]);
-
-const TreeItem = React.forwardRef(function TreeItem(
-  props: TreeItemProps & {
-    ContentProps?: { lastNestedChild?: boolean };
-  },
-  ref: React.Ref<HTMLLIElement>,
-) {
-  return <StyledTreeItem ContentComponent={CustomContent} {...props} ref={ref} />;
-});
-
 const code = `
-<TreeView
-  aria-label="file system navigator"
-  defaultExpanded={['1', '1.1', '1.2', '2', '2.3']}
-  sx={{ height: { xs: 260, sm: 460 }, overflowY: 'auto', p: 1 }}
->
-  <TreeItem itemId="1" label="Drive">
-    <TreeItem itemId="1.1" label="Backup">
-      <TreeItem
-        itemId="1.1.1"
-        label="Jan 2021.pdf"
-        ContentProps={{ lastNestedChild: true }}
-      />
-      <TreeItem
-        itemId="1.1.2"
-        label="Feb 2021.pdf"
-        ContentProps={{ lastNestedChild: true }}
-      />
-      <TreeItem
-        itemId="1.1.3"
-        label="Mar 2021.pdf"
-        ContentProps={{ lastNestedChild: true }}
-      />
-    </TreeItem>
-    <TreeItem itemId="1.2" label="Photos">
-      <TreeItem
-        itemId="1.2.1"
-        label="family.jpeg"
-        ContentProps={{ lastNestedChild: true }}
-      />
-      <TreeItem
-        itemId="1.2.2"
-        label="my_dogpng"
-        ContentProps={{ lastNestedChild: true }}
-      />
-    </TreeItem>
-  </TreeItem>
-  <TreeItem itemId="2" label="Favorite">
-    <TreeItem
-      itemId="2.1"
-      label="MUI_retreat_photo.jpg"
-      ContentProps={{ lastNestedChild: true }}
-    />
-    <TreeItem
-      itemId="2.2"
-      label="v6_secrets.mkv"
-      ContentProps={{ lastNestedChild: true }}
-    />
-    <TreeItem itemId="2.3" label="Other pictures">
-      <TreeItem
-        itemId="2.3.1"
-        label="my_avatar.jpg"
-        ContentProps={{ lastNestedChild: true }}
-      />
-    </TreeItem>
-  </TreeItem>
-</TreeView>`;
+<RichTreeView
+  items={ITEMS}
+  aria-label="file explorer"
+  defaultExpandedItems={['1', '1.1']}
+  defaultSelectedItems="1.1"
+  sx={{ height: 'fit-content', flexGrow: 1, overflowY: 'auto' }}
+  slots={{ item: CustomTreeItem }}
+/>`;
 
 export default function XTreeViewDemo() {
   return (
@@ -272,63 +286,14 @@ export default function XTreeViewDemo() {
             }),
           })}
         >
-          {/* <SimpleTreeView
-            aria-label="file system navigator"
-            defaultExpandedItems={['1', '1.1', '1.2']}
-            sx={{ height: { xs: 260, sm: '100%' }, overflowY: 'auto', p: 1 }}
-          >
-            <TreeItem itemId="1" label="Drive">
-              <TreeItem itemId="1.1" label="Backup">
-                <TreeItem
-                  itemId="1.1.1"
-                  label="Jan 2021.pdf"
-                  ContentProps={{ lastNestedChild: true }}
-                />
-                <TreeItem
-                  itemId="1.1.2"
-                  label="Feb 2021.pdf"
-                  ContentProps={{ lastNestedChild: true }}
-                />
-                <TreeItem
-                  itemId="1.1.3"
-                  label="Mar 2021.pdf"
-                  ContentProps={{ lastNestedChild: true }}
-                />
-              </TreeItem>
-              <TreeItem itemId="1.2" label="Photos">
-                <TreeItem
-                  itemId="1.2.1"
-                  label="family.jpeg"
-                  ContentProps={{ lastNestedChild: true }}
-                />
-                <TreeItem
-                  itemId="1.2.2"
-                  label="my_dogpng"
-                  ContentProps={{ lastNestedChild: true }}
-                />
-              </TreeItem>
-            </TreeItem>
-            <TreeItem itemId="2" label="Favorites">
-              <TreeItem
-                itemId="2.1"
-                label="MUI_retreat_photo.jpg"
-                ContentProps={{ lastNestedChild: true }}
-              />
-              <TreeItem
-                itemId="2.2"
-                label="v6_secrets.mkv"
-                ContentProps={{ lastNestedChild: true }}
-              />
-              <TreeItem itemId="2.3" label="Other pictures">
-                <TreeItem
-                  itemId="2.3.1"
-                  label="my_avatar.jpg"
-                  ContentProps={{ lastNestedChild: true }}
-                />
-              </TreeItem>
-            </TreeItem>
-          </SimpleTreeView> */}
-          <TreeViewDemo />
+          <RichTreeView
+            items={ITEMS}
+            aria-label="file explorer"
+            defaultExpandedItems={['1', '1.1', '2']}
+            defaultSelectedItems="1.1"
+            sx={{ height: 'fit-content', flexGrow: 1, overflowY: 'auto' }}
+            slots={{ item: CustomTreeItem }}
+          />
         </Paper>
       </Frame.Demo>
       <Frame.Info data-mui-color-scheme="dark" sx={{ maxHeight: 300, overflow: 'auto' }}>
