+++ docs/src/components/showcase/FolderTreeView.tsx
@@ -2,166 +2,248 @@
 /* eslint-disable jsx-a11y/click-events-have-key-events */
 import * as React from 'react';
 import clsx from 'clsx';
+import { animated, useSpring } from '@react-spring/web';
 import { styled } from '@mui/material/styles';
+import { TransitionProps } from '@mui/material/transitions';
 import Box from '@mui/material/Box';
-import { SimpleTreeView } from '@mui/x-tree-view/SimpleTreeView';
-import {
-  TreeItem,
-  useTreeItemState,
-  TreeItemProps,
-  TreeItemContentProps,
-} from '@mui/x-tree-view/TreeItem';
+import Collapse from '@mui/material/Collapse';
 import Typography from '@mui/material/Typography';
 import KeyboardArrowDownRounded from '@mui/icons-material/KeyboardArrowDownRounded';
 import KeyboardArrowUpRounded from '@mui/icons-material/KeyboardArrowUpRounded';
 import FolderRounded from '@mui/icons-material/FolderRounded';
+import { RichTreeView } from '@mui/x-tree-view/RichTreeView';
+import { treeItemClasses } from '@mui/x-tree-view/TreeItem';
+import {
+  unstable_useTreeItem2 as useTreeItem2,
+  UseTreeItem2Parameters,
+} from '@mui/x-tree-view/useTreeItem2';
+import {
+  TreeItem2Content,
+  TreeItem2IconContainer,
+  TreeItem2Label,
+  TreeItem2Root,
+} from '@mui/x-tree-view/TreeItem2';
+import { TreeItem2Icon } from '@mui/x-tree-view/TreeItem2Icon';
+import { TreeItem2Provider } from '@mui/x-tree-view/TreeItem2Provider';
+import { TreeViewBaseItem } from '@mui/x-tree-view/models';
 
-const CustomContent = React.forwardRef(function CustomContent(
-  props: TreeItemContentProps & { lastNestedChild?: boolean },
-  ref,
-) {
-  const {
-    lastNestedChild,
-    classes,
-    className,
-    label,
-    itemId,
-    icon: iconProp,
-    expansionIcon,
-    displayIcon,
-  } = props;
+type ExtendedTreeItemProps = {
+  id: string;
+  label: string;
+  color?: 'primary' | 'default';
+};
 
-  const {
-    disabled,
-    expanded,
-    selected,
-    focused,
-    handleExpansion,
-    handleSelection,
-    preventSelection,
-  } = useTreeItemState(itemId);
+const ITEMS: TreeViewBaseItem<ExtendedTreeItemProps>[] = [
+  {
+    id: '1',
+    label: 'src',
+    color: 'primary',
+    children: [
+      {
+        id: '1.1',
+        label: 'components',
+        children: [
+          { id: '1.1.1', label: 'Button.tsx' },
+          { id: '1.1.2', label: 'Drawer.tsx' },
+          { id: '1.1.3', label: 'Navbar.tsx' },
+          { id: '1.1.4', label: 'TreeView.tsx' },
+        ],
+      },
+      {
+        id: '1.2',
+        label: 'blocks',
+        children: [
+          { id: '1.2.1', label: 'SignupPage.tsx' },
+          {
+            id: '1.2.2',
+            label: 'PricingTable',
+            children: [
+              { id: '1.2.2.1', label: 'PaymentOptions.tsx' },
+              { id: '1.2.2.2', label: 'EarlyBirdDiscount.tsx' },
+            ],
+          },
+        ],
+      },
+    ],
+  },
+];
 
-  const icon = iconProp || expansionIcon || displayIcon;
-
-  const handleMouseDown = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    preventSelection(event);
-  };
-
-  const handleExpansionClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    handleExpansion(event);
-  };
-
-  const handleSelectionClick = (event: React.MouseEvent<HTMLDivElement, MouseEvent>) => {
-    handleSelection(event);
-  };
-
+function DotIcon() {
   return (
-    <div
-      className={clsx(className, classes.root, {
-        [classes.expanded]: expanded,
-        [classes.selected]: selected,
-        [classes.focused]: focused,
-        [classes.disabled]: disabled,
-      })}
-      // @ts-ignore
-      onClick={handleExpansionClick}
-      // @ts-ignore
-      onMouseDown={handleMouseDown}
-      // @ts-ignore
-      ref={ref as React.Ref<HTMLButtonElement>}
+    <Box
+      sx={{
+        width: 6,
+        height: 6,
+        borderRadius: '70%',
+        bgcolor: 'warning.main',
+        display: 'inline-block',
+        verticalAlign: 'middle',
+        zIndex: 1,
+        mx: 1,
+      }}
+    />
+  );
+}
+
+declare module 'react' {
+  interface CSSProperties {
+    '--tree-view-color'?: string;
+    '--tree-view-bg-color'?: string;
+  }
+}
+
+const StyledTreeItemRoot = styled(TreeItem2Root)(({ theme }) => ({
+  color: theme.palette.mode === 'light' ? theme.palette.grey[800] : theme.palette.grey[400],
+  position: 'relative',
+  [`& .${treeItemClasses.groupTransition}`]: {
+    paddingLeft: theme.spacing(0.5),
+  },
+})) as unknown as typeof TreeItem2Root;
+
+const CustomTreeItemContent = styled(TreeItem2Content)(({ theme }) => ({
+  borderRadius: theme.spacing(0.7),
+  marginBottom: theme.spacing(0.5),
+  marginTop: theme.spacing(0.5),
+  padding: theme.spacing(0.5),
+  [`& .${treeItemClasses.iconContainer}`]: {
+    marginRight: theme.spacing(1),
+  },
+
+  '&.Mui-expanded&::before': {
+    content: '""',
+    display: 'block',
+    position: 'absolute',
+    left: '16px',
+    top: '44px',
+    height: 'calc(100% - 48px)',
+    width: '1.5px',
+    backgroundColor:
+      theme.palette.mode === 'light'
+        ? (theme.vars || theme).palette.grey[100]
+        : (theme.vars || theme).palette.primaryDark[700],
+  },
+}));
+
+const AnimatedCollapse = animated(Collapse);
+
+function TransitionComponent(props: TransitionProps) {
+  const style = useSpring({
+    to: {
+      opacity: props.in ? 1 : 0,
+      transform: `translate3d(0,${props.in ? 0 : 20}px,0)`,
+      paddingLeft: 24,
+    },
+  });
+
+  return <AnimatedCollapse style={style} {...props} />;
+}
+
+interface CustomLabelProps {
+  children: React.ReactNode;
+  expandable?: boolean;
+  color: string;
+}
+
+function CustomLabel({ color, expandable, children, ...other }: CustomLabelProps) {
+  let Icon;
+  if (expandable) {
+    Icon = FolderRounded;
+  } else {
+    Icon = DotIcon;
+  }
+  return (
+    <TreeItem2Label
+      {...other}
+      sx={{
+        display: 'flex',
+        alignItems: 'center',
+      }}
     >
-      {lastNestedChild ? (
+      {Icon && (
         <Box
+          component={Icon}
+          className="labelIcon"
           sx={{
-            width: 6,
-            height: 6,
-            borderRadius: '50%',
-            bgcolor: 'warning.main',
-            display: 'inline-block',
-            verticalAlign: 'middle',
-            zIndex: 1,
-          }}
-        />
-      ) : (
-        <FolderRounded
-          sx={{
-            fontSize: 16,
-            color: itemId === '1' ? 'primary.main' : 'grey.600',
+            mr: 1,
+            fontSize: '1rem',
+            color,
           }}
         />
       )}
-      <Typography
-        onClick={handleSelectionClick}
-        component="div"
-        className={classes.label}
-        sx={{
-          '&&': {
-            color: lastNestedChild ? 'text.secondary' : 'text.primary',
-            fontWeight: lastNestedChild ? 400 : 500,
-          },
-        }}
-      >
-        {label}
-      </Typography>
-      {icon}
-    </div>
-  );
-});
 
-const StyledTreeItem = styled(TreeItem)(({ theme }) => [
-  {
-    '& .MuiTreeItem-content': {
-      border: 'none',
-      backgroundColor: 'transparent',
-      borderRadius: 5,
-      padding: theme.spacing(0.5),
-      textAlign: 'left',
-      position: 'relative',
-      zIndex: 1,
-    },
-    '& .MuiTreeItem-content .MuiTreeItem-label': {
-      paddingLeft: theme.spacing(1),
-    },
-    '& .MuiTreeItem-root': {
-      position: 'relative',
-      '&::before': {
-        content: '""',
-        display: 'block',
-        position: 'absolute',
-        left: -5,
-        height: '100%',
-        width: 1.5,
-        backgroundColor: (theme.vars || theme).palette.grey[200],
-      },
-    },
-    '& .MuiTreeItem-group': {
-      marginLeft: 0,
-      paddingLeft: theme.spacing(3),
-    },
-  },
-  theme.applyDarkStyles({
-    '& .MuiTreeItem-root': {
-      '&::before': {
-        backgroundColor: (theme.vars || theme).palette.primaryDark[600],
-      },
-    },
-    '& .MuiTreeItem-group': {
-      '& .MuiTreeItem-content': {
-        '&::before': {
-          backgroundColor: (theme.vars || theme).palette.primaryDark[600],
-        },
-      },
-    },
-  }),
-]);
+      <Typography
+        sx={(theme) => ({
+          fontWeight: expandable
+            ? theme.typography.fontWeightMedium
+            : theme.typography.fontWeightRegular,
+        })}
+        variant="body2"
+      >
+        {children}
+      </Typography>
+    </TreeItem2Label>
+  );
+}
+
+const isExpandable = (reactChildren: React.ReactNode): boolean => {
+  if (Array.isArray(reactChildren)) {
+    return reactChildren.length > 0 && reactChildren.some(isExpandable);
+  }
+  return Boolean(reactChildren);
+};
+
+interface CustomTreeItemProps
+  extends Omit<UseTreeItem2Parameters, 'rootRef'>,
+    Omit<React.HTMLAttributes<HTMLLIElement>, 'onFocus'> {}
 
 const CustomTreeItem = React.forwardRef(function CustomTreeItem(
-  props: TreeItemProps & {
-    ContentProps?: { lastNestedChild?: boolean };
-  },
+  props: CustomTreeItemProps,
   ref: React.Ref<HTMLLIElement>,
 ) {
-  return <StyledTreeItem ContentComponent={CustomContent} {...props} ref={ref} />;
+  const { id, itemId, label, disabled, children, ...other } = props;
+
+  const {
+    getRootProps,
+    getContentProps,
+    getIconContainerProps,
+    getLabelProps,
+    getGroupTransitionProps,
+    status,
+    publicAPI,
+  } = useTreeItem2({ id, itemId, children, label, disabled, rootRef: ref });
+
+  const item = publicAPI.getItem(itemId);
+  const expandable = isExpandable(children);
+
+  return (
+    <TreeItem2Provider itemId={itemId}>
+      <StyledTreeItemRoot {...getRootProps(other)}>
+        <CustomTreeItemContent
+          {...getContentProps({
+            className: clsx('content', {
+              'Mui-expanded': status.expanded,
+              'Mui-selected': status.selected,
+              'Mui-focused': status.focused,
+              'Mui-disabled': status.disabled,
+            }),
+          })}
+        >
+          <CustomLabel
+            {...getLabelProps({
+              expandable,
+              color: item?.color === 'primary' ? 'primary.main' : 'grey.600',
+            })}
+          />{' '}
+          {expandable && (
+            <TreeItem2IconContainer {...getIconContainerProps()}>
+              <TreeItem2Icon status={status} />
+            </TreeItem2IconContainer>
+          )}
+        </CustomTreeItemContent>
+        {children && <TransitionComponent {...getGroupTransitionProps()} />}{' '}
+      </StyledTreeItemRoot>
+    </TreeItem2Provider>
+  );
 });
 function CustomEndIcon() {
   return <div style={{ width: 24 }} />;
@@ -175,49 +257,20 @@ function CustomCollapseIcon() {
   return <KeyboardArrowUpRounded sx={{ fontSize: 16, color: 'primary.main' }} />;
 }
 
-export default function FolderTreeView() {
+export default function TreeViewDemo() {
   return (
-    <SimpleTreeView
-      aria-label="folder"
-      defaultExpandedItems={['1', '2', '5', '7']}
+    <RichTreeView
+      items={ITEMS}
+      aria-label="file explorer"
+      defaultExpandedItems={['1', '1.1', '1.2', '1.2.2']}
+      defaultSelectedItems="1.1"
+      sx={{ height: 'fit-content', flexGrow: 1, overflowY: 'auto', p: 1 }}
       slots={{
+        item: CustomTreeItem,
         endIcon: CustomEndIcon,
         expandIcon: CustomExpandIcon,
         collapseIcon: CustomCollapseIcon,
       }}
-      sx={{ p: 1, overflowY: 'auto' }}
-    >
-      <CustomTreeItem itemId="1" label="src">
-        <CustomTreeItem itemId="2" label="components">
-          <CustomTreeItem itemId="3" label="Button.tsx" ContentProps={{ lastNestedChild: true }} />
-          <CustomTreeItem itemId="4" label="Drawer.tsx" ContentProps={{ lastNestedChild: true }} />
-          <CustomTreeItem itemId="5" label="Navbar.tsx" ContentProps={{ lastNestedChild: true }} />
-          <CustomTreeItem
-            itemId="6"
-            label="TreeView.tsx"
-            ContentProps={{ lastNestedChild: true }}
-          />
-        </CustomTreeItem>
-        <CustomTreeItem itemId="7" label="blocks">
-          <CustomTreeItem
-            itemId="8"
-            label="SignUpPage.tsx"
-            ContentProps={{ lastNestedChild: true }}
-          />
-          <CustomTreeItem itemId="9" label="PricingTable.tsx">
-            <CustomTreeItem
-              itemId="10"
-              label="PaymentOptions.tsx"
-              ContentProps={{ lastNestedChild: true }}
-            />
-            <CustomTreeItem
-              itemId="11"
-              label="EarlyBirdDiscount.tsx"
-              ContentProps={{ lastNestedChild: true }}
-            />
-          </CustomTreeItem>
-        </CustomTreeItem>
-      </CustomTreeItem>
-    </SimpleTreeView>
+    />
   );
 }
